Remove derived CGNR and CGNE specialziations for CG/CA-CG/CG3: we now have single implementations of these, which use a solver factor to call the correct CG method
